+++ airticketopen-20230117/core/Models/SearchResponseBody.cs
@@ -443,6 +443,20 @@ namespace AlibabaCloud.SDK.AirticketOpen20230117.Models
 
                 }
 
+                [NameInMap("solution_attribute")]
+                [Validation(Required=false)]
+                public SearchResponseBodyDataSolutionListSolutionAttribute SolutionAttribute { get; set; }
+                public class SearchResponseBodyDataSolutionListSolutionAttribute : TeaModel {
+                    /// <summary>
+                    /// <b>Example:</b>
+                    /// <para>1</para>
+                    /// </summary>
+                    [NameInMap("supply_source_type")]
+                    [Validation(Required=false)]
+                    public string SupplySourceType { get; set; }
+
+                }
+
                 /// <summary>
                 /// <para>solution ID</para>
                 ///